requirements.txt
# Optional: replace with actual MCP client lib when available
dspy>=0.2.0
fastapi>=0.110.0
langchain-openai>=0.1.0
langgraph>=0.4.0
openai>=1.14.0
uvicorn[standard]>=0.29.0
jsonschema>=4.24.0
opentelemetry-instrumentation-fastapi>=0.45b0
opentelemetry-sdk>=1.23.0
opentelemetry-api>=1.23.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
langfuse>=3.1.3